View Details
This needs to change a bit, there is a problem with downloading files,
also this page should be its own page somewhere in the system and we
need to add settings for the other modules that need to view resources.
The download files needs some work for this to work, will address it
tomorrow.